platform, instrument, and location keywords
This update looks for the following keywords and their vocabularies in global attributes and adds them to the ncISO output:

platform
platform_vocabulary
instrument
instrument_vocabulary
locations
locations_vocabulary

Note that platform and instrument are already "suggested" global attributes in the ACDD-1.3 convention. I have proposed adding locations and locations_vocabulary into the next ACDD version with favorable initial response from ESIP. Note that I previously modified ncISO to look for location keywords in THREDDSMetadata. This update continues to look there but gives precedence to the global attributes if present, especially because locations_vocabulary is not captured in THREDDSMetadata.